requirements.txt
Flask==2.3.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.5.2
SQLAlchemy==2.0.19
mysql-connector-python==8.1.0
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
Werkzeug==2.3.7
